Split admin to multiple more focuses blueprints
